refactor: move two repl classes
