+++ cli-console/src/main/java/org/aya/cli/repl/jline/JlineRepl.java
@@ -4,9 +4,9 @@ package org.aya.cli.repl.jline;
 
 import com.intellij.lexer.FlexLexer;
 import kala.collection.immutable.ImmutableSeq;
+import org.aya.cli.interactive.ReplConfig;
 import org.aya.cli.render.RenderOptions;
 import org.aya.cli.repl.AyaRepl;
-import org.aya.cli.repl.ReplConfig;
 import org.aya.cli.repl.gk.GKReplLexer;
 import org.aya.generic.util.AyaHome;
 import org.aya.parser.AyaParserDefinitionBase;
