+++ cli-impl/src/main/java/module-info.java
@@ -8,6 +8,7 @@ module aya.cli.impl {
   requires com.google.gson;
   requires org.commonmark;
 
+  exports org.aya.cli.interactive;
   exports org.aya.cli.library.incremental;
   exports org.aya.cli.library.json;
   exports org.aya.cli.library.source;
