+++ cli-impl/src/main/java/org/aya/cli/interactive/ReplContext.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl;
+package org.aya.cli.interactive;
 
 import kala.collection.Seq;
 import kala.collection.immutable.ImmutableSeq;
@@ -24,11 +24,10 @@ public final class ReplContext extends PhysicalModuleContext implements RepoLike
 
   @Override
   public void addGlobal(
-    @NotNull ImmutableSeq<String> modName,
-    @NotNull String name,
-    Stmt.@NotNull Accessibility accessibility,
-    @NotNull AnyVar ref,
-    @NotNull SourcePos sourcePos) {
+    @NotNull ImmutableSeq<String> modName, @NotNull String name,
+    Stmt.@NotNull Accessibility accessibility, @NotNull AnyVar ref,
+    @NotNull SourcePos sourcePos
+  ) {
     definitions.getOrPut(name, MutableHashMap::of).set(modName, ref);
     if (accessibility == Stmt.Accessibility.Public) {
       exports.get(TOP_LEVEL_MOD_NAME).exportAnyway(name, ref);
@@ -38,20 +37,18 @@ public final class ReplContext extends PhysicalModuleContext implements RepoLike
   @Override
   public void importModule(
     Stmt.@NotNull Accessibility accessibility,
-    @NotNull SourcePos sourcePos,
-    ImmutableSeq<String> componentName,
-    ModuleExport mod) {
+    @NotNull SourcePos sourcePos, ImmutableSeq<String> componentName,
+    ModuleExport mod
+  ) {
     modules.put(componentName, mod);
     if (accessibility == Stmt.Accessibility.Public) exports.set(componentName, mod);
   }
 
-  @Override
-  public @NotNull ReplContext derive(@NotNull Seq<@NotNull String> extraName) {
+  @Override public @NotNull ReplContext derive(@NotNull Seq<@NotNull String> extraName) {
     return new ReplContext(this, this.moduleName().concat(extraName));
   }
 
-  @Override
-  public @NotNull ReplContext derive(@NotNull String extraName) {
+  @Override public @NotNull ReplContext derive(@NotNull String extraName) {
     return new ReplContext(this, this.moduleName().appended(extraName));
   }
