+++ cli-impl/src/main/java/org/aya/cli/interactive/ReplShapeFactory.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl;
+package org.aya.cli.interactive;
 
 import org.aya.cli.utils.RepoLike;
 import org.aya.core.repr.AyaShape;
