+++ cli-impl/src/test/java/org/aya/cli/ReplTestBase.java
@@ -6,8 +6,8 @@ import kala.collection.immutable.ImmutableSeq;
 import kala.control.Option;
 import kala.tuple.Tuple;
 import kala.tuple.Tuple2;
+import org.aya.cli.interactive.ReplConfig;
 import org.aya.cli.repl.AyaRepl;
-import org.aya.cli.repl.ReplConfig;
 import org.aya.repl.IO;
 import org.intellij.lang.annotations.Language;
 import org.jetbrains.annotations.NotNull;
